--- doc/source/client — копия.rst
@@ -1,7 +1,0 @@
-client — копия module
-=====================
-
-.. automodule:: client — копия
-   :members:
-   :undoc-members:
-   :show-inheritance:
