+++ doc/source/client.rst
@@ -1,5 +1,67 @@
-client module
-=============
+client package
+==============
+
+Submodules
+----------
+
+client.add\_contact\_dialog module
+----------------------------------
+
+.. automodule:: client.add_contact_dialog
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+client.client\_database module
+------------------------------
+
+.. automodule:: client.client_database
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+client.client\_thread module
+----------------------------
+
+.. automodule:: client.client_thread
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+client.del\_contact\_dialog module
+----------------------------------
+
+.. automodule:: client.del_contact_dialog
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+client.main\_win module
+-----------------------
+
+.. automodule:: client.main_win
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+client.main\_win\_converted module
+----------------------------------
+
+.. automodule:: client.main_win_converted
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+client.start\_dialog module
+---------------------------
+
+.. automodule:: client.start_dialog
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
 
 .. automodule:: client
    :members:
