+++ doc/source/clnt.rst
@@ -1,0 +1,7 @@
+clnt module
+===========
+
+.. automodule:: clnt
+   :members:
+   :undoc-members:
+   :show-inheritance:
