+++ doc/source/conf.py
@@ -3,31 +3,26 @@
 # For the full list of built-in configuration values, see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
 
+import os
+import sys
+path = os.path.dirname(os.path.dirname(os.path.abspath('.')))
+sys.path.insert(0, path)
+
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-import sys
-from pathlib import Path
-sys.path.append(Path().absolute().parent.parent)
-
-project = 'PyQt_ORM_geekbrains_course'
+project = 'Geelbrains_PyQt_ORM'
 copyright = '2022, KolesnikIvan'
 author = 'KolesnikIvan'
-release = '001'
+release = '002'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = [
-    'sphinx.ext.duration',
-    'sphinx.ext.doctest',
-    'sphinx.ext.autodoc',
-]
+extensions = ['sphinx.ext.autodoc',]
 
 templates_path = ['_templates']
 exclude_patterns = []
-
-
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
