--- doc/source/server — копия.rst
@@ -1,77 +1,0 @@
-server — копия package
-======================
-
-Submodules
-----------
-
-server — копия.core module
---------------------------
-
-.. automodule:: server — копия.core
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-server — копия.core — копия module
-----------------------------------
-
-.. automodule:: server — копия.core — копия
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-server — копия.srv\_db module
------------------------------
-
-.. automodule:: server — копия.srv_db
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-server — копия.srv\_gui module
-------------------------------
-
-.. automodule:: server — копия.srv_gui
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-server — копия.win\_add\_user module
-------------------------------------
-
-.. automodule:: server — копия.win_add_user
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-server — копия.win\_config module
----------------------------------
-
-.. automodule:: server — копия.win_config
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-server — копия.win\_hist module
--------------------------------
-
-.. automodule:: server — копия.win_hist
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-server — копия.win\_rem\_user module
-------------------------------------
-
-.. automodule:: server — копия.win_rem_user
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: server — копия
-   :members:
-   :undoc-members:
-   :show-inheritance:
