+++ doc/source/server.rst
@@ -1,5 +1,75 @@
-server module
-=============
+server package
+==============
+
+Submodules
+----------
+
+server.core module
+------------------
+
+.. automodule:: server.core
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+server.core — копия module
+--------------------------
+
+.. automodule:: server.core — копия
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+server.srv\_db module
+---------------------
+
+.. automodule:: server.srv_db
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+server.srv\_gui module
+----------------------
+
+.. automodule:: server.srv_gui
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+server.win\_add\_user module
+----------------------------
+
+.. automodule:: server.win_add_user
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+server.win\_config module
+-------------------------
+
+.. automodule:: server.win_config
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+server.win\_hist module
+-----------------------
+
+.. automodule:: server.win_hist
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+server.win\_rem\_user module
+----------------------------
+
+.. automodule:: server.win_rem_user
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
 
 .. automodule:: server
    :members:
