+++ doc/source/srvr.rst
@@ -1,0 +1,7 @@
+srvr module
+===========
+
+.. automodule:: srvr
+   :members:
+   :undoc-members:
+   :show-inheritance:
